Restarting on net flop
cjdns likes to randomly go into "Network unreachable" state for some
reason, so delay sending until it stops throwing a hissy fit.

It might possibly be my cjdns dying for no reason in the background...
still good to add robustness in the face of such things.

Was going to restart the whole shebang, then realized ENETUNREACH
doesn't kill the socket so can just keep trying to send to it. Total
restart of the meshchat UDP socket might be useful for uh... something
that kills the socket? Maybe?

I've done nothing productive all day!